iterate: rename commands to spec
