+++ turtle/src/test/kotlin/com/lordcodes/turtle/commands/LsCommandTest.kt
@@ -2,6 +2,9 @@
 
 package com.lordcodes.turtle.commands
 
+import com.lordcodes.turtle.specs.CommandOption
+import com.lordcodes.turtle.specs.FileSpec
+import com.lordcodes.turtle.specs.command
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import java.io.File
@@ -11,14 +14,14 @@ class LsCommandTest {
     @Test
     fun `ls without arguments`() {
         val expected = command("ls", ".")
-        val actual = FileCommands.ls()
+        val actual = FileSpec.ls()
         assertEquals(expected, actual)
     }
 
     @Test
     fun `ls with short arguments`() {
         val expected = command("ls", "-a", "-b", "-c", "/etc")
-        val actual = FileCommands.ls(
+        val actual = FileSpec.ls(
             files = listOf(File("/etc")),
             args = listOf("-a", "-b", "-c"),
         )
@@ -28,7 +31,7 @@ class LsCommandTest {
     @Test
     fun `ls with long arguments`() {
         val expected = command("ls", "--all", "--colour", "--blocks", "420", "/etc")
-        val actual = FileCommands.ls(
+        val actual = FileSpec.ls(
             files = listOf(File("/etc")),
         ) {
             listOf(all, colour, blocks.withValue("420"))
@@ -38,7 +41,7 @@ class LsCommandTest {
 
     @Test
     fun `ls with warnings`() {
-        FileCommands.ls() {
+        FileSpec.ls() {
             listOf(CommandOption("--whatever"), CommandOption("--invalid"))
         }
         /*
